added react navigation and created basic play page
